sched/fair: Don't needlessly migrate a lone task to a higher capacity CPU
When a CPU has only one running CFS task and there's a higher capacity CPU
that's idle, that lone task may be migrated to the higher capacity CPU just
because of RT and IRQ load.

If the CPU running the lone CFS task has sufficient capacity for the task,
then let it run that task. Migrating the task up to a higher capacity CPU
causes that CPU to be kicked out of idle and degrades energy efficiency
without an appreciable performance improvement. The load balancer will take
care of migrating the task anyway if it becomes a misfit, so this heuristic
isn't needed.

Change-Id: I1683f25deda7bbb6c16752c716d8d0db8571da9f
Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Richard Raya <rdxzv.dev@gmail.com>